+++ JiraPS/Public/Get-JiraStatus.ps1
@@ -1,0 +1,62 @@
+﻿function Get-JiraStatus {
+    # .ExternalHelp ..\JiraPS-help.xml
+    [CmdletBinding( DefaultParameterSetName = '_All' )]
+    param(
+        [Parameter( Position = 0, Mandatory, ValueFromPipeline, ParameterSetName = '_Search' )]
+        [Alias('IdOrName')]
+        [String[]]
+        $Status,
+
+        [Parameter()]
+        [System.Management.Automation.PSCredential]
+        [System.Management.Automation.Credential()]
+        $Credential = [System.Management.Automation.PSCredential]::Empty
+    )
+
+    begin {
+        Write-Verbose "[$($MyInvocation.MyCommand.Name)] Function started"
+
+        $server = Get-JiraConfigServer -ErrorAction Stop
+
+        $resourceURi = "$server/rest/api/latest/status{0}"
+    }
+
+    process {
+        Write-DebugMessage "[$($MyInvocation.MyCommand.Name)] ParameterSetName: $($PsCmdlet.ParameterSetName)"
+        Write-DebugMessage "[$($MyInvocation.MyCommand.Name)] PSBoundParameters: $($PSBoundParameters | Out-String)"
+
+        switch ($PSCmdlet.ParameterSetName) {
+            '_All' {
+                $parameter = @{
+                    URI        = $resourceURi -f ""
+                    Method     = "GET"
+                    Credential = $Credential
+                }
+                Write-Debug "[$($MyInvocation.MyCommand.Name)] Invoking JiraMethod with `$parameter"
+                $result = Invoke-JiraMethod @parameter
+
+                Write-Output (ConvertTo-JiraStatus -InputObject $result)
+            }
+            '_Search' {
+                foreach ($_status in $Status) {
+                    Write-Verbose "[$($MyInvocation.MyCommand.Name)] Processing [$_status]"
+                    Write-Debug "[$($MyInvocation.MyCommand.Name)] Processing `$_status [$_status]"
+
+                    $parameter = @{
+                        URI        = $resourceURi -f "/$_status"
+                        Method     = "GET"
+                        Credential = $Credential
+                    }
+                    Write-Debug "[$($MyInvocation.MyCommand.Name)] Invoking JiraMethod with `$parameter"
+                    $result = Invoke-JiraMethod @parameter
+
+                    Write-Output (ConvertTo-JiraStatus -InputObject $result)
+                }
+            }
+        }
+    }
+
+    end {
+        Write-Verbose "[$($MyInvocation.MyCommand.Name)] Complete"
+    }
+}